minor
git-svn-id: file:///home/fnevgeny/svn/fac/trunk/fac@234 ad7935cb-1c1c-0410-8b58-bafadb17b910

+++ faclib/crm.h
@@ -69,7 +69,7 @@ typedef struct _ION_ {
 typedef struct _IONIZED_ {
   int nele;
   char symbol[4];
-  int atom;
+  double atom;
   char *dbfiles[NDB];
   int nionized;
   int *ionized_map[2];
